feat(vault): remove legacy VaultView/AccountsView/NotificationsView
The web vault's unlocked UI is now fully served by the shared
AccountSettingsView (sidebar + per-account tabs + Vault Settings) wired to
real routes, so the old tabbed VaultView shell and its AccountsView /
NotificationsView panels are dead code. Remove them and their tests.

The one VaultView responsibility not already covered — the "Desktop app
connected" success banner (vaultConnectionSuccessMessage) — moves into
AccountSettingsView. A new AccountSettingsView.test.tsx covers routed
account selection, sidebar clicks, and the connection banner. The obsolete
hash-based account-selection tests are dropped in favor of real path routes.

Note: the legacy drag-to-reorder accounts affordance is not carried over to
the new sidebar; revisit if account ordering is needed.
