--- vault/src/frontend/views/VaultView.test.tsx
@@ -1,271 +1,0 @@
-import {act, cleanup, fireEvent, render, screen, waitFor} from '@testing-library/react'
-import {afterEach, describe, expect, mock, test} from 'bun:test'
-import * as blobs from '@shm/shared/blobs'
-import {Account, Profile} from '@shm/shared/client/.generated/documents/v3alpha/documents_pb'
-import {MemoryRouter} from 'react-router-dom'
-import {StoreContext, createStore} from '@/frontend/store'
-import {APIError} from '@/frontend/api-client'
-import {createMockBlockstore, createSuccessMockClient} from '@/frontend/test-utils'
-import {VaultView} from './VaultView'
-
-describe('VaultView', () => {
-  afterEach(() => {
-    cleanup()
-  })
-
-  function createVaultStore(
-    names = ['Alice'],
-    notificationServerUrl = '',
-    createTime = 123456789,
-    populateProfiles = true,
-  ) {
-    const store = createStore(
-      createSuccessMockClient(),
-      createMockBlockstore(),
-      'https://daemon.example.com',
-      notificationServerUrl,
-    )
-    store.state.sessionChecked = true
-    store.state.session = {
-      authenticated: true,
-      relyingPartyOrigin: 'https://example.com',
-      email: 'test@example.com',
-    }
-    store.state.decryptedDEK = new Uint8Array(32)
-    const principals: string[] = []
-    store.state.vaultData = {
-      version: 2,
-      accounts: names.map((name, index) => {
-        const seed = new Uint8Array(32).fill(index + 1)
-        const principal = blobs.principalToString(blobs.nobleKeyPairFromSeed(seed).principal)
-        principals.push(principal)
-        if (populateProfiles) {
-          store.state.profiles[principal] = {name}
-        }
-
-        return {
-          seed,
-          createTime,
-          delegations: [],
-        }
-      }),
-    }
-    store.state.selectedAccountIndex = principals.length > 0 ? 0 : -1
-
-    return {store, principals}
-  }
-
-  test('selects the account from the hash route after unlock', async () => {
-    const originalFetch = global.fetch
-    const originalPath = `${window.location.pathname}${window.location.search}${window.location.hash}`
-    const fetchMock = mock(async () => {
-      return new Response(
-        JSON.stringify({
-          accountId: 'account-1',
-          email: null,
-          verifiedTime: null,
-          verificationSendTime: null,
-          verificationExpired: false,
-          isRegistered: false,
-        }),
-        {
-          status: 200,
-          headers: {'Content-Type': 'application/json'},
-        },
-      )
-    })
-    global.fetch = fetchMock as unknown as typeof fetch
-    const {store, principals} = createVaultStore(['Alice', 'Bob'], 'https://notify.example.com')
-    history.replaceState(null, '', '/vault')
-    window.location.hash = `/a/${encodeURIComponent(principals[1]!)}`
-
-    try {
-      render(
-        <MemoryRouter>
-          <StoreContext.Provider value={store}>
-            <VaultView />
-          </StoreContext.Provider>
-        </MemoryRouter>,
-      )
-
-      await waitFor(() => {
-        expect(store.state.selectedAccountIndex).toBe(1)
-      })
-      expect(window.location.hash).toBe(`#/a/${encodeURIComponent(principals[1]!)}`)
-    } finally {
-      global.fetch = originalFetch
-      history.replaceState(null, '', originalPath)
-    }
-  })
-
-  test('updates the hash when the selected account changes', async () => {
-    const originalFetch = global.fetch
-    const originalPath = `${window.location.pathname}${window.location.search}${window.location.hash}`
-    const fetchMock = mock(async () => {
-      return new Response(
-        JSON.stringify({
-          accountId: 'account-1',
-          email: null,
-          verifiedTime: null,
-          verificationSendTime: null,
-          verificationExpired: false,
-          isRegistered: false,
-        }),
-        {
-          status: 200,
-          headers: {'Content-Type': 'application/json'},
-        },
-      )
-    })
-    global.fetch = fetchMock as unknown as typeof fetch
-    const {store, principals} = createVaultStore(['Alice', 'Bob'], 'https://notify.example.com')
-    history.replaceState(null, '', '/vault')
-
-    try {
-      render(
-        <MemoryRouter>
-          <StoreContext.Provider value={store}>
-            <VaultView />
-          </StoreContext.Provider>
-        </MemoryRouter>,
-      )
-
-      await waitFor(() => {
-        expect(window.location.hash).toBe(`#/a/${encodeURIComponent(principals[0]!)}`)
-      })
-
-      await act(async () => {
-        store.actions.selectAccount(1)
-      })
-
-      await waitFor(() => {
-        expect(store.state.selectedAccountIndex).toBe(1)
-        expect(window.location.hash).toBe(`#/a/${encodeURIComponent(principals[1]!)}`)
-      })
-    } finally {
-      global.fetch = originalFetch
-      history.replaceState(null, '', originalPath)
-    }
-  })
-
-  test('refreshes missing account profile names when the picker opens', async () => {
-    const originalPath = `${window.location.pathname}${window.location.search}${window.location.hash}`
-    const seed = new Uint8Array(32).fill(1)
-    const principal = blobs.principalToString(blobs.nobleKeyPairFromSeed(seed).principal)
-    let getAccountCalls = 0
-    const getAccount = mock(async () => {
-      getAccountCalls += 1
-      if (getAccountCalls === 1) {
-        throw new APIError('Account not found', 404)
-      }
-
-      return new Account({
-        id: principal,
-        profile: new Profile({
-          name: 'Alice',
-        }),
-      })
-    })
-    const store = createStore(
-      createSuccessMockClient({getAccount}),
-      createMockBlockstore(),
-      'https://daemon.example.com',
-      'https://notify.example.com',
-    )
-    store.state.sessionChecked = true
-    store.state.session = {
-      authenticated: true,
-      relyingPartyOrigin: 'https://example.com',
-      email: 'test@example.com',
-    }
-    store.state.decryptedDEK = new Uint8Array(32)
-    store.state.vaultData = {
-      version: 2,
-      accounts: [
-        {
-          seed,
-          createTime: 123456789,
-          delegations: [],
-        },
-      ],
-    }
-    store.state.selectedAccountIndex = 0
-    history.replaceState(null, '', '/vault')
-
-    try {
-      render(
-        <MemoryRouter>
-          <StoreContext.Provider value={store}>
-            <VaultView />
-          </StoreContext.Provider>
-        </MemoryRouter>,
-      )
-
-      await waitFor(() => {
-        expect(getAccountCalls).toBe(1)
-      })
-
-      const pickerButton = screen.getAllByRole('button').find((button) => button.textContent?.includes(principal))
-      expect(pickerButton).toBeDefined()
-      fireEvent.click(pickerButton!)
-
-      await waitFor(() => {
-        expect(getAccountCalls).toBe(2)
-        expect(screen.getAllByText('Alice').length).toBeGreaterThan(0)
-      })
-    } finally {
-      history.replaceState(null, '', originalPath)
-    }
-  })
-
-  test('shows and dismisses the desktop connection success banner', async () => {
-    const originalFetch = global.fetch
-    const originalPath = `${window.location.pathname}${window.location.search}${window.location.hash}`
-    const fetchMock = mock(async () => {
-      return new Response(
-        JSON.stringify({
-          accountId: 'account-1',
-          email: null,
-          verifiedTime: null,
-          verificationSendTime: null,
-          verificationExpired: false,
-          isRegistered: false,
-        }),
-        {
-          status: 200,
-          headers: {'Content-Type': 'application/json'},
-        },
-      )
-    })
-    global.fetch = fetchMock as unknown as typeof fetch
-    const {store} = createVaultStore(['Alice'], 'https://notify.example.com')
-    store.state.vaultConnectionSuccessMessage =
-      'Your Seed desktop app has been linked with this remote vault successfully.'
-    history.replaceState(null, '', '/vault')
-
-    try {
-      render(
-        <MemoryRouter>
-          <StoreContext.Provider value={store}>
-            <VaultView />
-          </StoreContext.Provider>
-        </MemoryRouter>,
-      )
-
-      expect(screen.getByText('Desktop app connected')).toBeTruthy()
-      expect(
-        screen.getByText('Your Seed desktop app has been linked with this remote vault successfully.'),
-      ).toBeTruthy()
-
-      fireEvent.click(screen.getByRole('button', {name: 'Dismiss desktop app connected message'}))
-
-      await waitFor(() => {
-        expect(store.state.vaultConnectionSuccessMessage).toBe('')
-      })
-      expect(screen.queryByText('Desktop app connected')).toBeNull()
-    } finally {
-      global.fetch = originalFetch
-      history.replaceState(null, '', originalPath)
-    }
-  })
-})
